% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_pcap.R
\name{dataFrame}
\alias{dataFrame}
\title{dati e frame inviati tcp}
\usage{
dataFrame(pcapName, fluxType)
}
\arguments{
\item{pcapName:}{nome del file pcap}

\item{fluxType:}{protocollo del flusso}
}
\value{
list in cui list[0]= numero frame list[1]= byte inviati
}
\description{
dati e frame inviati tcp
}
\examples{
library(TesinaTIR)
#dati e frame inviati tcp
pcapName<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
res<-dataFrame(pcapName,'tcp')
tcp_frame<-res[[1]]
tcp_data<-res[[2]]
print('totale frame tcp:')
print(tcp_frame)
print('totale dati tcp:')
print(tcp_data)
#dati e frame inviati udp
res<-dataFrame(pcapName,'udp')
udp_frame<-res[[1]]
udp_data<-res[[2]]
print('totale frame udp:')
print(udp_frame)
print('totale dati udp:')
print(udp_data)
}

