% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_count.R
\name{fluxCount}
\alias{fluxCount}
\title{conteggio flussi tcp}
\usage{
fluxCount(pcapName, fluxType)
}
\arguments{
\item{pcapName:}{nome del pcap}

\item{fluxType:}{protocollo del flusso}
}
\value{
numero di flussi del protocollo fluxType
}
\description{
conteggio flussi tcp
}
\examples{
library(TesinaTIR)
pcapName<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
n_tcp_flux<-fluxCount(pcapName, 'tcp')
n_udp_flux<-fluxCount(pcapName, 'udp')
}

