% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_parser.R
\name{get_argv}
\alias{get_argv}
\title{parser dei parametri passati dalla command line}
\usage{
get_argv(description, arguments)
}
\arguments{
\item{description}{descrizione dello script}

\item{arguments}{argomenti accettati dallo script in un array con i seguenti elementi:\cr 
arguments[i]=c(argument,description,default_value) \cr
dove c() è il comando di r per creare un array}
}
\value{
argv validato
}
\description{
parametri di input dello script di default: \cr
-i: input file, default: dump.pcap \cr
-o: output directory, default: ./output \cr
-h: help\cr
Tutti gli altri parametri devono essere specificati con il parametro arguments.
}
\examples{
library(TesinaTIR)
description<-"
descrizione dello script
"
#argomenti aggiuntivi
arguments<-matrix(c("--bin_size",
                   "dimensione del bin per il campionamento di campTcpUdp","1"), 
                 ncol=3,nrow=1)
#argv contiene i parametri passati da command line
argv<-get_argv(description,arguments)
}
\seealso{
\code{\link{argparser}} la funzione utilizza questa libreria
}

