% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conv_durations_TcpUdp.R
\name{get_conv_durations_TcpUdp}
\alias{get_conv_durations_TcpUdp}
\title{durata di tutti i flussi udp/tcp:}
\usage{
get_conv_durations_TcpUdp(filename, protocol)
}
\arguments{
\item{filename}{file da leggere}

\item{protocol}{protocollo}
}
\value{
array che contiene la durata di tutti i flussi del protocollo specificato
}
\description{
durata di tutti i flussi udp/tcp:
}
\examples{
library(TesinaTIR)
filename<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
#durate tcp
durations_tcp<-get_conv_durations_TcpUdp(filename,'tcp')
#durate udp 
durations_udp<-get_conv_durations_TcpUdp(filename,'udp')
}

