% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conv_timeRelative_frameLen.R
\name{get_conv_timeRelative_frameLen}
\alias{get_conv_timeRelative_frameLen}
\title{reperisce le tracce relative ad un flusso e ne calcola il tempo relativo rispetto all'inizio del flusso.}
\usage{
get_conv_timeRelative_frameLen(stream_id, protocol, filename)
}
\arguments{
\item{stream_id}{id del flusso}

\item{protocol}{protocollo}

\item{filename}{nome del file da leggere}
}
\value{
array con i tempi relativi all'inizio del flusso delle tracce del flusso identificato da stream_id
}
\description{
reperisce le tracce relative ad un flusso e ne calcola il tempo relativo rispetto all'inizio del flusso.
}
\examples{
library(TesinaTIR)
stream<-1
protocol<-'tcp'
filename<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
trace_tcp<-get_conv_timeRelative_frameLen(stream,protocol,filename)
stream<-1
protocol<-'udp'
filename<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
trace_udp<-get_conv_timeRelative_frameLen(stream,protocol,filename)
}

