% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_stream_id.R
\name{get_max_stream_id}
\alias{get_max_stream_id}
\title{stream id massimo del pcap \cr}
\usage{
get_max_stream_id(filename, protocol)
}
\arguments{
\item{filename:}{nome del file di input}

\item{protocol:}{protocollo}
}
\value{
stream_id massimo del pcap
}
\description{
stream id va da 0 a 'MAX(tcp.stream)tcp.stream'
}
\examples{
library(TesinaTIR)
filename<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
max_stream_id_tcp<-get_max_stream_id(filename,'tcp')
max_stream_id_udp<-get_max_stream_id(filename,'udp')
}

