% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TcpConv_correct_fwh.R
\name{get_TcpConv_correct_fwh}
\alias{get_TcpConv_correct_fwh}
\title{flussi terminati con fin \cr}
\usage{
get_TcpConv_correct_fwh(filename, streams_id)
}
\arguments{
\item{filename}{nome del file pcap}

\item{streams_id}{vettore che continene gli id di tutti gli stream del pcap}
}
\value{
array che contiene per ogni stream tcp una variabile booleana: \cr
FALSE: flusso non terminato da fin
TRUE: flusso terminato da fin
}
\description{
Questa funzione si occupa di calcolare se i flussi in input sono stati
terminati da due fin o meno.\cr
comando di prova \cr
tshark -r GazzettaDiModena_20161120.pcap  -z io,stat,0,'COUNT(tcp.flags.fin)tcp.flags.fin and tcp.flags.fin==1 and tcp.stream==11' -q \cr
}
\examples{
library(TesinaTIR)
filename<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
#tempo massimo relativo del pcap
tmax<-tmaxPcap(filename)
#estrazione flussi sottoforma di stream id
#stream id va da 0 a 'MAX(tcp.stream)tcp.stream'
#stream id massimo
max_stream_id<-get_max_stream_id(filename,'tcp')
#genero tutti gli stream id
streams_id<-get_streams_id(max_stream_id)
#terminazione flusso con fin
fin<-get_TcpConv_correct_fwh(filename,streams_id)
}

