% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TcpConv_correct_twh.R
\name{get_TcpConv_correct_twh}
\alias{get_TcpConv_correct_twh}
\title{presenza three way handshake nei flussi}
\usage{
get_TcpConv_correct_twh(filename, streams_id)
}
\arguments{
\item{filename}{nome del file pcap}

\item{streams_id}{id degli stream tcp del pcap}
}
\value{
array che contiene per ogni stream tcp una variabile booleana: \cr
#' FALSE: twh non eseguito
#' TRUE: twh eseguito
}
\description{
Questa funzione calcola la presenza o meno del twh nei flussi di input. \cr
il filtro di wireshark utilizzato è composto dalle seguenti parti: \cr
* tcp.flags.syn==1 cattura i primi due pacchetti \cr
* tcp.seq==1 and tcp.ack==1  and tcp.analysis.initial_rtt cattura il terzo pacchetto, \cr
  perchè wireshark calcola tcp.analysis.initial_rtt ogni volta che viene instaurata \cr
  una connessione tcp ed è applicato a tutte le tracce successive all'instaurazione \cr
  della connessione \cr
comando per testare questo filtro \cr
tshark -r  GazzettaDiModena_20161120.pcap -Y 'tcp.flags.syn==1 or (tcp.seq==1 and tcp.ack==1  and tcp.analysis.initial_rtt) and tcp.stream==11' \cr
prendo le due tracce iniziali: syn, syn-ack \cr
tshark -r  GazzettaDiModena_20161120.pcap -Y 'tcp.flags.syn==1 and tcp.seq==0 and tcp.stream==11' \cr
prendo la terza traccia: ack \cr
tshark -r  GazzettaDiModena_20161120.pcap -Y 'tcp.seq==1 and tcp.ack==1  and tcp.analysis.initial_rtt and tcp.stream==11' \cr
se ho almeno due tracce per le tracce iniziali e almento una traccia per qella finale allora il three way handshake è instaurato con successo \cr
}
\examples{
library(TesinaTIR)
filename<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
#tempo massimo relativo del pcap
tmax<-tmaxPcap(filename)

#estrazione flussi sottoforma di stream id
#stream id va da 0 a 'MAX(tcp.stream)tcp.stream'
#stream id massimo
max_stream_id<-get_max_stream_id(filename,'tcp')

#genero tutti gli stream id
streams_id<-get_streams_id(max_stream_id)
#calcolo i flussi che hanno eseguito il three way hanshake 
three_way_handshake<-get_TcpConv_correct_twh(filename,streams_id)
}

