% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_bin.R
\name{normalize_bin}
\alias{normalize_bin}
\title{funzione che normalizza il valore dei bins.}
\usage{
normalize_bin(bins_count)
}
\arguments{
\item{bins_count}{vettore che contiene i conteggio dei bins}
}
\value{
array con i valori dei bin normalizzati
}
\description{
funzione che normalizza il valore dei bins.
}
\examples{
library(TesinaTIR)
bins_count<-c(8,1,1)
normalize_bin(bins_count) 
}

