% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtt_time.R
\name{rtt_time}
\alias{rtt_time}
\title{metriche ack rtt del tcp sull'intervallo temporale [tmin,tmax]}
\usage{
rtt_time(tmin, tmax, step, filename)
}
\arguments{
\item{tmin}{tempo minimo}

\item{tmax}{tempo massimo}

\item{step}{dimensione dei bin}

\item{filename}{nome del file}
}
\value{
dataframe che contiene le seguenti colonne \cr
l_time:valore temporale del bin \cr
l_max: tempo massimo nel bin \cr
l_min: valore minimo nel bin \cr
l_avg: valore medio nel bin \cr
l_var; varianza nel bin
}
\description{
Questa funzione calcola media,varianza, valore massimo e minimo di ack rtt.\cr
}
\examples{
library(TesinaTIR)
tmax<-5
bin_size<-1
filename<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
rtt_tcp_intervals<-rtt_time(0,tmax,bin_size,filename)
}
\seealso{
\code{\link[TesinaTIR]{tcp_ack_rtt}} la funzione dipende da questa funzione
}

