% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_values_TcpUdp.R
\name{time_values_TcpUdp}
\alias{time_values_TcpUdp}
\title{andamento flussi tcp e udp nell'intervallo temporale [tmin,tmax]}
\usage{
time_values_TcpUdp(pcapName, tmin, tmax, step)
}
\arguments{
\item{pcapName}{file pcap da leggere}

\item{tmin}{tempo minimo}

\item{tmax}{tempo massimo}

\item{step}{dimensione bins}
}
\value{
dataframe composto dalle seguenti colonne: \cr
l_time: valori temporali dei bins \cr
l_tcp_frame: numero di frame tcp \cr
l_udp_frame: numero di frame udp \cr
l_udp_data: byte inviati da flussi udp \cr
l_tcp_data: byte inviati da flussi tcp \cr
l_tcp_duplicate_ack: numero di duplicate ack del tcp
}
\description{
andamento flussi tcp e udp nell'intervallo temporale [tmin,tmax]
}
\examples{
library(TesinaTIR)
pcapName<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
tmax<-5
step_size<-1
campTcpUdp<-time_values_TcpUdp(pcapName,0,tmax,step_size)
}

