% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmax_pcap.R
\name{tmaxPcap}
\alias{tmaxPcap}
\title{tempo massimo relativo del pcap}
\usage{
tmaxPcap(pcapName)
}
\arguments{
\item{pcapName}{nome del file pcap}
}
\description{
tempo massimo relativo del pcap
}
\examples{
library(TesinaTIR)
pcapName<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
tmax<-tmaxPcap(pcapName)
}
\seealso{
\code{\link[stringr]{str_trim}} la funzione dipende da questa funzione
}

