% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendConvTcpUdp.R
\name{trendConvTcpUdp}
\alias{trendConvTcpUdp}
\title{andamento numero flussi tcp e udp nel tempo}
\usage{
trendConvTcpUdp(pcapName, tmin, tmax, step)
}
\arguments{
\item{pcapName}{file pcap da leggere}

\item{tmin}{tempo minimo}

\item{tmax}{tempo massimo}

\item{step}{dimensione bins}
}
\value{
dataframe composto dalle seguenti colonne: \cr
l_time: valori temporali dei bins \cr
l_tcp_flux: numero di frame tcp \cr
l_udp_flux: numero di frame tcp \cr
}
\description{
andamento numero flussi tcp e udp nel tempo
}
\examples{
library(TesinaTIR)
pcapName<-system.file("extdata", "dump.pcap", package = "TesinaTIR")
tmax<-5
step_size<-1
n_flux<-trendConvTcpUdp(pcapName,0,tmax,step_size)
}

